Beginning restructuring by new template.
