+++ .editorconfig
@@ -4,7 +4,10 @@ root = true
 [*]
 charset = utf-8
 end_of_line = lf
-indent_size = 4
+indent_size = 2
 indent_style = space
+
+[*.php]
+indent_size = 4
 insert_final_newline = true
 trim_trailing_whitespace = true
