+++ composer.json
@@ -1,6 +1,6 @@
 {
-  "name": "clickalicious/memcached.php",
-  "description": "Memcached.php - Plain vanilla PHP Memcached client with full support of Memcached protocol.",
+  "name": "clickalicious/memcached-php",
+  "description": "Plain vanilla PHP Memcached client library with full support of Memcached ASCII protocol.",
   "type": "library",
   "authors": [
     {
@@ -13,19 +13,25 @@
   "minimum-stability": "stable",
   "prefer-stable": true,
   "require": {
-    "php": ">=5.4.0",
-    "gpupo/cache": "^1.2"
+    "php": ">=5.6.0",
+    "roave/security-advisories": "dev-master as 1.0.x-dev",
+    "gpupo/cache": "^1.3"
   },
   "require-dev": {
-    "satooshi/php-coveralls": "^1.0",
-    "phpunit/phpunit": "^4.8"
+      "phpunit/phpunit": "^5.7",
+      "codacy/coverage": "^1.0.5",
+      "symfony/var-dumper": "^3.2"
   },
   "autoload": {
-      "psr-4": {
-          "Clickalicious\\": "src/Clickalicious/"
-      }
+    "psr-4": {
+      "Clickalicious\\": "src/Clickalicious/"
+    }
   },
-  "include-path": ["src/"],
+  "autoload-dev": {
+    "psr-4": {
+      "Clickalicious\\": "src/Clickalicious/"
+    }
+  },
   "keywords": [
     "Memcached",
     "Client",
@@ -35,15 +41,16 @@
     "inmemory",
     "nosql"
   ],
-  "homepage": "https://github.com/clickalicious/Memcached.php",
-  "license": "BSD-3-Clause",
+  "homepage": "https://github.com/clickalicious/memcached-php",
+  "license": "MIT",
   "support": {
     "email": "opensource@clickalicious.de",
-    "issues": "https://github.com/clickalicious/Memcached.php/issues",
-    "wiki": "https://github.com/clickalicious/Memcached.php/wiki"
+    "issues": "https://github.com/clickalicious/memcached-php/issues",
+    "wiki": "https://github.com/clickalicious/memcached-php/wiki",
+    "tests": "https://clickalicious.github.io/memcached-php"
   },
   "config" : {
     "bin-dir" : "bin",
-    "preferred-install": "dist"
+    "preferred-install": "source"
   }
 }
