+++ phpunit.xml.dist
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/4.5/phpunit.xsd"
+>
+  <testsuites>
+    <testsuite name="clickalicious Test Suite">
+      <directory suffix="Test.php">./tests</directory>
+    </testsuite>
+  </testsuites>
+
+  <filter>
+    <whitelist>
+      <directory>./src</directory>
+      <exclude>
+        <directory>./bin</directory>
+        <directory>./docs</directory>
+        <directory>./vendor</directory>
+        <directory>./tests</directory>
+      </exclude>
+    </whitelist>
+  </filter>
+</phpunit>
