--- src/Clickalicious/Memcached/Cache.php
@@ -1,153 +1,0 @@
-<?php
-/* vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4: */
-
-namespace Clickalicious\Memcached;
-
-/**
- * Memcached.php
- *
- * Cache.php - PSR compatbile Cache-Client() based on Client() - Plain vanilla PHP Memcached client with full
- * support of Memcached protocol.
- *
- *
- * PHP versions 5.3
- *
- * LICENSE:
- * Memcached.php - Plain vanilla PHP Memcached client with full support of Memcached protocol.
- *
- * Copyright (c) 2014 - 2015, Benjamin Carl
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * - Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * - Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * - Neither the name of Memcached.php nor the names of its
- * contributors may be used to endorse or promote products derived from
- * this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * Please feel free to contact us via e-mail: opensource@clickalicious.de
- *
- * @category   Clickalicious
- * @package    Clickalicious_Memcached
- * @subpackage Clickalicious_Memcached_Cache
- * @author     Benjamin Carl <opensource@clickalicious.de>
- * @copyright  2014 - 2015 Benjamin Carl
- * @license    http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @version    Git: $Id$
- * @link       https://github.com/clickalicious/Memcached.php
- */
-
-use Psr\Cache\CacheItemInterface;
-
-/**
- * Memcached.php
- *
- * PSR compatbile Cache-Client() based on Client() - Plain vanilla PHP Memcached client with full
- * support of Memcached protocol.
- *
- * @category   Clickalicious
- * @package    Clickalicious_Memcached
- * @subpackage Clickalicious_Memcached_Cache
- * @author     Benjamin Carl <opensource@clickalicious.de>
- * @copyright  2014 - 2015 Benjamin Carl
- * @license    http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @version    Git: $Id$
- * @link       https://github.com/clickalicious/Memcached.php
- */
-class Cache extends Client implements CacheItemInterface
-{
-    /**
-     * Returns the key for the current cache item.
-     *
-     * The key is loaded by the Implementing Library, but should be available to
-     * the higher level callers when needed.
-     *
-     * @return string
-     *   The key string for this cache item.
-     */
-    public function getKey()
-    {
-
-    }
-
-    /**
-     * Confirms if the cache item lookup resulted in a cache hit.
-     *
-     * Note: This method MUST NOT have a race condition between calling isHit()
-     * and calling get().
-     *
-     * @return boolean
-     *   True if the request resulted in a cache hit.  False otherwise.
-     */
-    public function isHit()
-    {
-
-    }
-
-    /**
-     * Confirms if the cache item exists in the cache.
-     *
-     * Note: This method MAY avoid retrieving the cached value for performance
-     * reasons, which could result in a race condition between exists() and get().
-     * To avoid that potential race condition use isHit() instead.
-     *
-     * @return boolean
-     *  True if item exists in the cache, false otherwise.
-     */
-    public function exists()
-    {
-
-    }
-
-    /**
-     * Sets the expiration for this cache item.
-     *
-     * @param int|\DateTime $ttl
-     *   - If an integer is passed, it is interpreted as the number of seconds
-     *     after which the item MUST be considered expired.
-     *   - If a DateTime object is passed, it is interpreted as the point in
-     *     time after which the item MUST be considered expired.
-     *   - If null is passed, a default value MAY be used. If none is set,
-     *     the value should be stored permanently or for as long as the
-     *     implementation allows.
-     *
-     * @return static
-     *   The called object.
-     */
-    public function setExpiration($ttl = null)
-    {
-
-    }
-
-    /**
-     * Returns the expiration time of a not-yet-expired cache item.
-     *
-     * If this cache item is a Cache Miss, this method MAY return the time at
-     * which the item expired or the current time if that is not available.
-     *
-     * @return \DateTime
-     *   The timestamp at which this cache item will expire.
-     */
-    public function getExpiration()
-    {
-
-    }
-}
