+++ src/Clickalicious/Memcached/Php/Cache.php
@@ -1,0 +1,126 @@
+<?php
+
+/**
+ * (The MIT license)
+ * Copyright 2017 clickalicious, Benjamin Carl
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+ * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+namespace Clickalicious\Memcached\Php;
+
+use Psr\Cache\CacheItemInterface;
+
+/**
+ * Class Cache
+ *
+ * @package Clickalicious\Memcached\Php
+ * @author  Benjamin Carl <opensource@clickalicious.de>
+ */
+class Cache extends Client implements CacheItemInterface
+{
+    /**
+     * Returns the key for the current cache item.
+     *
+     * The key is loaded by the Implementing Library, but should be available to
+     * the higher level callers when needed.
+     *
+     * @return string
+     *   The key string for this cache item.
+     */
+    public function getKey()
+    {
+
+    }
+
+    /**
+     * Confirms if the cache item lookup resulted in a cache hit.
+     *
+     * Note: This method MUST NOT have a race condition between calling isHit()
+     * and calling get().
+     *
+     * @return boolean
+     *   True if the request resulted in a cache hit.  False otherwise.
+     */
+    public function isHit()
+    {
+
+    }
+
+    /**
+     * Confirms if the cache item exists in the cache.
+     *
+     * Note: This method MAY avoid retrieving the cached value for performance
+     * reasons, which could result in a race condition between exists() and get().
+     * To avoid that potential race condition use isHit() instead.
+     *
+     * @return boolean
+     *  True if item exists in the cache, false otherwise.
+     */
+    public function exists()
+    {
+
+    }
+
+    /**
+     * Sets the expiration for this cache item.
+     *
+     * @param int|\DateTime $ttl
+     *   - If an integer is passed, it is interpreted as the number of seconds
+     *     after which the item MUST be considered expired.
+     *   - If a DateTime object is passed, it is interpreted as the point in
+     *     time after which the item MUST be considered expired.
+     *   - If null is passed, a default value MAY be used. If none is set,
+     *     the value should be stored permanently or for as long as the
+     *     implementation allows.
+     *
+     * @return static
+     *   The called object.
+     */
+    public function setExpiration($ttl = null)
+    {
+
+    }
+
+    /**
+     * Returns the expiration time of a not-yet-expired cache item.
+     *
+     * If this cache item is a Cache Miss, this method MAY return the time at
+     * which the item expired or the current time if that is not available.
+     *
+     * @return \DateTime
+     *   The timestamp at which this cache item will expire.
+     */
+    public function getExpiration()
+    {
+
+    }
+
+    public function expiresAfter($time)
+    {
+        // TODO: Implement expiresAfter() method.
+    }
+
+    public function expiresAt($expiration)
+    {
+        // TODO: Implement expiresAt() method.
+    }
+}
