+++ src/Clickalicious/Memcached/Php/Compression/Zlib.php
@@ -1,0 +1,69 @@
+<?php
+
+/**
+ * (The MIT license)
+ * Copyright 2017 clickalicious, Benjamin Carl
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+ * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+namespace Clickalicious\Memcached\Php\Compression;
+
+/**
+ * Class Zlib
+ *
+ * Zlib.php - ZLIB-compression.
+ *
+ * ZLIB-compression interface to builtin PHP functionality.
+ *
+ * @package Clickalicious\Memcached\Php\Compression
+ * @author  Benjamin Carl <opensource@clickalicious.de>
+ */
+class Zlib implements CompressionInterface
+{
+    /**
+     * Compresses a buffer.
+     *
+     * @param string $buffer The buffer to compress
+     *
+     * @author Benjamin Carl <opensource@clickalicious.de>
+     * @return string The compressed input
+     * @access protected
+     */
+    public function compress($buffer)
+    {
+        return gzcompress($buffer);
+    }
+
+    /**
+     * Decompresses a buffer.
+     *
+     * @param string $buffer The buffer to decompress
+     *
+     * @author Benjamin Carl <opensource@clickalicious.de>
+     * @return string The decompressed input
+     * @access protected
+     */
+    public function decompress($buffer)
+    {
+        return gzuncompress($buffer);
+    }
+}
