+++ tests/Clickalicious/Memcached/Php/UtilTest.php
@@ -1,77 +1,42 @@
 <?php
-/* vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4: */
 
 /**
- * Memcached.php
+ * (The MIT license)
+ * Copyright 2017 clickalicious, Benjamin Carl
  *
- * UtilTest.php - Unit tests for util functionality.
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
  *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
  *
- * PHP versions 5.3
- *
- * LICENSE:
- * Memcached.php - Plain vanilla PHP Memcached client with full support of Memcached protocol.
- *
- * Copyright (c) 2014 - 2015, Benjamin Carl
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * - Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * - Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * - Neither the name of Memcached.php nor the names of its
- * contributors may be used to endorse or promote products derived from
- * this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * Please feel free to contact us via e-mail: opensource@clickalicious.de
- *
- * @category   Clickalicious
- * @package    Clickalicious_Memcached
- * @subpackage Clickalicious_Memcached_Tests
- * @author     Benjamin Carl <opensource@clickalicious.de>
- * @copyright  2014 - 2015 Benjamin Carl
- * @license    http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @version    Git: $Id$
- * @link       https://github.com/clickalicious/Memcached.php
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+ * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
-require_once CLICKALICIOUS_MEMCACHED_BASE_PATH . 'Clickalicious/Memcached/Util.php';
+namespace Clickalicious\Memcached\Php;
 
 /**
- * Memcached.php
+ * Class UtilTest
  *
- * Unit tests for util functionality.
- *
- * @category   Clickalicious
- * @package    Clickalicious_Memcached
- * @subpackage Clickalicious_Memcached_Tests
- * @author     Benjamin Carl <opensource@clickalicious.de>
- * @copyright  2014 - 2015 Benjamin Carl
- * @license    http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
- * @version    Git: $Id$
- * @link       https://github.com/clickalicious/Memcached.php
+ * @package Clickalicious\Memcached\Php
+ * @author  Benjamin Carl <opensource@clickalicious.de>
  */
-class UtilTest extends PHPUnit_Framework_TestCase
+class UtilTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * The PHP version running on.
+     * PHP version running on.
      *
      * @var string
      * @access protected
@@ -154,7 +119,7 @@ class UtilTest extends PHPUnit_Framework_TestCase
             [3] => Peter
         )
         */
-        $data = \Clickalicious\Memcached\array_column_emulation($this->data, 'first_name');
+        $data = \Clickalicious\Memcached\Php\array_column_emulation($this->data, 'first_name');
 
         $this->assertContains('John',  $data);
         $this->assertContains('Sally', $data);
@@ -185,7 +150,7 @@ class UtilTest extends PHPUnit_Framework_TestCase
             [5623] => Doe
         )
          */
-        $data = \Clickalicious\Memcached\array_column_emulation($this->data, 'first_name', 'id');
+        $data = \Clickalicious\Memcached\Php\array_column_emulation($this->data, 'first_name', 'id');
 
         $this->assertContains('John',  $data);
         $this->assertContains('Sally', $data);
@@ -216,7 +181,7 @@ class UtilTest extends PHPUnit_Framework_TestCase
             [5623] => Doe
         )
         */
-        $data = \Clickalicious\Memcached\array_column_emulation($this->data, 'foo');
+        $data = \Clickalicious\Memcached\Php\array_column_emulation($this->data, 'foo');
         $this->assertArrayHasKey(0, $data);
         $this->assertArrayHasKey(1, $data);
         $this->assertArrayHasKey(2, $data);
@@ -241,7 +206,7 @@ class UtilTest extends PHPUnit_Framework_TestCase
             [5623] => Doe
         )
          */
-        $data = \Clickalicious\Memcached\array_column_emulation($this->data, 'first_name', 1);
+        $data = \Clickalicious\Memcached\Php\array_column_emulation($this->data, 'first_name', 1);
         $this->assertArrayHasKey(0, $data);
         $this->assertArrayHasKey(1, $data);
         $this->assertArrayHasKey(2, $data);
@@ -266,7 +231,7 @@ class UtilTest extends PHPUnit_Framework_TestCase
             [5623] => Doe
         )
         */
-        $data = \Clickalicious\Memcached\array_column_emulation($this->data, 'first_name', 'foo');
+        $data = \Clickalicious\Memcached\Php\array_column_emulation($this->data, 'first_name', 'foo');
         $this->assertArrayHasKey(0, $data);
         $this->assertArrayHasKey(1, $data);
         $this->assertArrayHasKey(2, $data);
@@ -283,7 +248,7 @@ class UtilTest extends PHPUnit_Framework_TestCase
      */
     public function testArrayColumnEmulationErrorHandlingWrongSecondArgument()
     {
-        \Clickalicious\Memcached\array_column_emulation($this->data, new stdClass());
+        \Clickalicious\Memcached\Php\array_column_emulation($this->data, new stdClass());
     }
 
     /**
@@ -296,7 +261,7 @@ class UtilTest extends PHPUnit_Framework_TestCase
      */
     public function testArrayColumnEmulationErrorHandlingWrongThirdArgument()
     {
-        \Clickalicious\Memcached\array_column_emulation($this->data, 'first_name', new stdClass());
+        \Clickalicious\Memcached\Php\array_column_emulation($this->data, 'first_name', new stdClass());
     }
 
     /**
@@ -318,7 +283,7 @@ class UtilTest extends PHPUnit_Framework_TestCase
                 [3] => Peter
             )
              */
-            $data = \Clickalicious\Memcached\array_column($this->data, 'first_name');
+            $data = \Clickalicious\Memcached\Php\array_column($this->data, 'first_name');
 
             $this->assertContains('John',  $data);
             $this->assertContains('Sally', $data);
@@ -348,30 +313,30 @@ class UtilTest extends PHPUnit_Framework_TestCase
     public function testBoolvalEmulation()
     {
         $data = true;
-        $this->assertTrue(\Clickalicious\Memcached\boolval_emulation($data));
+        $this->assertTrue(\Clickalicious\Memcached\Php\boolval_emulation($data));
 
         $data = 1;
-        $this->assertTrue(\Clickalicious\Memcached\boolval_emulation($data));
+        $this->assertTrue(\Clickalicious\Memcached\Php\boolval_emulation($data));
 
         $data = 'TRUE';
-        $this->assertTrue(\Clickalicious\Memcached\boolval_emulation($data));
+        $this->assertTrue(\Clickalicious\Memcached\Php\boolval_emulation($data));
 
         $data = 'YES';
-        $this->assertTrue(\Clickalicious\Memcached\boolval_emulation($data));
+        $this->assertTrue(\Clickalicious\Memcached\Php\boolval_emulation($data));
 
         $data = 'Y';
-        $this->assertTrue(\Clickalicious\Memcached\boolval_emulation($data));
+        $this->assertTrue(\Clickalicious\Memcached\Php\boolval_emulation($data));
 
         $data = 'ON';
-        $this->assertTrue(\Clickalicious\Memcached\boolval_emulation($data));
+        $this->assertTrue(\Clickalicious\Memcached\Php\boolval_emulation($data));
 
         $data = '1';
-        $this->assertTrue(\Clickalicious\Memcached\boolval_emulation($data));
+        $this->assertTrue(\Clickalicious\Memcached\Php\boolval_emulation($data));
 
         $data = '-1';
-        $this->assertTrue(\Clickalicious\Memcached\boolval_emulation($data));
+        $this->assertTrue(\Clickalicious\Memcached\Php\boolval_emulation($data));
 
         $data = false;
-        $this->assertFalse(\Clickalicious\Memcached\boolval_emulation($data));
+        $this->assertFalse(\Clickalicious\Memcached\Php\boolval_emulation($data));
     }
 }
