--- tests/phpunit.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/4.8/phpunit.xsd"
-         bootstrap="../src/Clickalicious/Memcached/Bootstrap.php"
-         backupGlobals="false"
-         backupStaticAttributes="false"
-         cacheTokens="false"
-         colors="true"
-         stopOnError="true"
-         stopOnFailure="true"
-         verbose="false"
-         testdox="true"
-         timeoutForSmallTests="1"
-         timeoutForMediumTests="10"
-         timeoutForLargeTests="60"
->
-    <testsuites>
-        <testsuite name="Memcached.php">
-            <directory>.</directory>
-        </testsuite>
-    </testsuites>
-
-    <filter>
-        <whitelist>
-            <directory>../src</directory>
-            <exclude>
-                <directory>../vendor</directory>
-                <directory>../src/Clickalicious/Memcached/Autoloader.php</directory>
-                <directory>../src/Clickalicious/Memcached/Bootstrap.php</directory>
-                <directory>../src/Clickalicious/Memcached/Cache.php</directory>
-                <directory>../src/Clickalicious/Memcached/Exception.php</directory>
-                <directory>../src/Clickalicious/Memcached/Compression</directory>
-            </exclude>
-        </whitelist>
-    </filter>
-</phpunit>
